+ refactoring

+++ src/main/java/dbservice/service/CategoryServiceImpl.java
@@ -1,16 +1,22 @@
 package dbservice.service;
 
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
+import com.google.gson.Gson;
+import com.google.gson.JsonElement;
+import com.google.gson.JsonObject;
+import com.google.gson.JsonParser;
 import dbservice.converter.CategoryConverter;
 import dbservice.dao.CategoryDao;
+import dbservice.dao.ProductDao;
 import dbservice.dto.CategoryDto;
 import dbservice.entity.Category;
+import dbservice.entity.Product;
+import dbservice.result.StatusResult;
+import dbservice.result.Message;
+import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -23,11 +29,22 @@ public class CategoryServiceImpl implements CategoryService {
     @Autowired
     private CategoryConverter categoryConverter;
 
+    @Autowired
+    private ProductDao productDao;
+
+    private static final Logger logger = Logger.getLogger(CategoryServiceImpl.class);
+
     @Override
     @Transactional
     public CategoryDto getById(long id){
         return categoryConverter.convertToDto(categoryDao.getById(id));
 
+    }
+
+    @Override
+    @Transactional
+    public CategoryDto getByName(String name){
+        return categoryConverter.convertToDto(categoryDao.getByName(name));
     }
 
     @Override
@@ -42,30 +59,65 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
 
+
     @Override
     @Transactional
-    public void updateFromJson (String categoryJson) throws IOException {
-        ObjectMapper mapper = new ObjectMapper();
-        JsonNode rootNode = mapper.readTree(categoryJson);
-        System.out.println(rootNode);
-        Long id = rootNode.get("id").asLong();
-        String name = rootNode.get("name").asText();
+    public String updateCategory (String categoryJson){
+        JsonElement jElement = new JsonParser().parse(categoryJson);
+        JsonObject jObject = jElement.getAsJsonObject();
+        Long id = jObject.get("id").getAsLong();
+        String name = jObject.get("name").getAsString();
+        CategoryDto category = getByName(name);
+        CategoryDto updatedCategory = getById(id);
+        Gson gson = new Gson();
+        if (category == null || category.getId().equals(id)){
+            updatedCategory.setCategoryName(name);
+            update(updatedCategory);
+            String message = StatusResult.CATEGORY_SUCCESS_UPDATE.getFormatMessage(name);
+            logger.info(message);
+            return gson.toJson(new Message(StatusResult.CATEGORY_SUCCESS_UPDATE, message));
+        }
+        String message = StatusResult.CATEGORY_ALREADY_EXIST.getFormatMessage(name);
+        logger.info(message);
+        //JsonObject object = new JsonObject();
+        //object.addProperty("resultType", StatusResult.CATEGORY_ALREADY_EXIST.name());
+        //object.addProperty("message", message);
+        //object.addProperty("oldName", updatedCategory.getCategoryName());
+        //return gson.toJson(object);
+        Message jsonMessage = new Message(StatusResult.CATEGORY_ALREADY_EXIST, message);
+        jsonMessage.getDataMap().put("oldName", updatedCategory.getCategoryName());
+        return gson.toJson(jsonMessage);
+    }
 
+    @Override
+    @Transactional
+    public StatusResult add(CategoryDto categoryDto){
+        CategoryDto category = getByName(categoryDto.getCategoryName());
+        if (category == null) {
+            categoryDao.add(categoryConverter.convertToEntity(categoryDto));
+            logger.info(String.format(StatusResult.CATEGORY_SUCCESS_CREATE.getMessage(), categoryDto.getCategoryName()));
+            return StatusResult.CATEGORY_SUCCESS_CREATE;
+        }
+        logger.info(String.format(StatusResult.CATEGORY_ALREADY_EXIST.getMessage(), categoryDto.getCategoryName()));
+        return StatusResult.CATEGORY_ALREADY_EXIST;
+
+    }
+
+    @Override
+    @Transactional
+    public String deleteById(long id){
+        Gson gson = new Gson();
         CategoryDto categoryDto = getById(id);
-        categoryDto.setCategoryName(name);
-        update(categoryDto);
-    }
-
-    @Override
-    @Transactional
-    public void add(CategoryDto categoryDto){
-        categoryDao.add(categoryConverter.convertToEntity(categoryDto));
-    }
-
-    @Override
-    @Transactional
-    public void deleteById(long id){
-        categoryDao.deleteById(id);
+        List<Product> products = productDao.getByCategory(categoryConverter.convertToEntity(categoryDto));
+        if (products.size() == 0) {
+            categoryDao.deleteById(id);
+            String message = StatusResult.CATEGORY_SUCCESS_DELETE.getFormatMessage(categoryDto.getCategoryName());
+            logger.info(message);
+            return gson.toJson(new Message(StatusResult.CATEGORY_SUCCESS_DELETE, message));
+        }
+        String message = StatusResult.CATEGORY_FAIL_DELETE.getFormatMessage(categoryDto.getCategoryName());
+        logger.info(message);
+        return gson.toJson(new Message(StatusResult.CATEGORY_FAIL_DELETE, message));
     }
 
     @Override
